Moved ITransactionNode from shared/src to front-end/src/shared/interfaces
