+++ front-end/src/renderer/pages/Transactions/components/History.vue
@@ -44,7 +44,7 @@ import DateTimeString from '@renderer/components/ui/DateTimeString.vue';
 import TransactionId from '@renderer/components/ui/TransactionId.vue';
 import { useRouter } from 'vue-router';
 import useTableQueryState from '@renderer/composables/useTableQueryState.ts';
-import { TransactionNodeCollection } from '../../../../../../shared/src/ITransactionNode.ts';
+import { TransactionNodeCollection } from '@shared/interfaces/ITransactionNode.ts';
 import { AppCache } from '@renderer/caches/AppCache.ts';
 
 const HISTORY_SORT_URL_VALUES = [
