+++ front-end/src/shared/interfaces/ITransactionNode.ts
